Completes modularisation of types and utils
Moves shared types and utils to dedicated directories,
improving project structure and maintainability.

Updates ignore files to reflect the new directory structure
and exclude unnecessary files from builds and version control.

Updates documentation to reflect the new locations of types and utils.
