+++ .npmignore
@@ -18,16 +18,25 @@ docs/experimental/
 node_modules/
 # Allow dist/extension.cjs and dist/assets for runtime
 !dist/extension.cjs
+!dist/extension.js
+!dist/cli.js
+!dist/commandHandler.js
+!dist/index.js
+!dist/mcpServer.js
+!dist/mcpServerCli.js
+!dist/memoryBank.js
+!dist/memoryBankCore.js
+!dist/memoryBankServiceCore.js
+!dist/types.js
+!dist/webviewManager.js
 !dist/assets/**
 !dist/lib/rules/**
 !dist/webview/**
-!dist/cli.js
-!dist/commandHandler.js
-!dist/mcpServer*
-!dist/memoryBank*
-!dist/extension.js
-!dist/types.ts
-dist/*.js.map
+!dist/core/**
+!dist/mcp/**
+!dist/types/**
+!dist/utils/**
+!dist/*.js.map
 
 # Exclude environment and log files
 .env
@@ -47,3 +56,7 @@ npm-debug.log
 
 # Allow esbuild.js for npm scripts
 !esbuild.js
+
+# Exclude test output and system files
+dist/test/
+dist/.DS_Store
