+++ docs/wip/TODO.md
@@ -6,17 +6,19 @@
 -   [x] Implement robust MCP server with error handling
 -   [x] Add webview UI for memory bank actions
 -   [x] Migrate flat memory bank to modular structure
+-   [x] Add MCP CLI/stdio entrypoint for Cursor-first operation
+-   [x] Self-healing memory bank and rules management
+-   [x] MCP tools exposed and robust
+-   [~] Refactor extension to remove Express and use Cursor/VS Code APIs for all communication. (in progress)
+-   [~] User-configurable log levels for Output Channel (in progress: basic output visible, needs verbose logging and interactivity)
+-   [~] Webview error and event reporting to Output Channel (in progress)
 -   [ ] Add version control integration for memory bank files
 -   [ ] Implement remote/cloud memory bank support
--   [~] Refactor extension to remove Express and use Cursor/VS Code APIs for all communication. (in progress)
 -   [ ] Ensure all communication and MCP tool logic is Cursor-first (Cursor compatibility is the top priority; VS Code compatibility is a bonus).
--   [~] Add a dedicated Output Channel for AI Memory extension logs using the Cursor/VS Code API. (in progress: basic output visible, needs verbose logging and interactivity)
--   [ ] Review and analyse MCP and Developer Tools logs in `logs/` (especially `2025-05-09 15:19:50.log` and matching `vscode-app-*.log` files) for startup, error, and runtime issues.
+-   [ ] Review and analyse MCP and Developer Tools logs in `logs/` for startup, error, and runtime issues.
 -   [x] Rebuild and repackage the extension after refactoring (ignore files reviewed and fixed, 0.1.5 is new stable baseline).
 -   [ ] Test activation, command registration, and MCP tool operation in Cursor (and optionally VS Code).
 -   [ ] Document findings and next steps in `docs/ROADMAP.md`.
--   [~] User-configurable log levels for Output Channel (see docs/experimental/detailed-output-with-webview.md)
--   [~] Webview error and event reporting to Output Channel (see docs/experimental/detailed-output-with-webview.md)
 -   [ ] Replace `Initialize Memory Bank` with `/memory init` command
 -   [ ] Add command: 'AI Memory: Create Memory Bank Rule' to create/restore memory-bank.mdc from template or rules source
     -   [ ] Implement user prompt to overwrite if file exists (see cursor-rules-service.ts)
@@ -24,6 +26,7 @@
 ## UI/UX
 
 -   [x] Webview: Initialise/Update Memory Bank buttons
+-   [x] Webview UI overhaul (React/Tailwind, status, repair, management)
 -   [ ] Fix "Initialise Memory Bank" button
     -   [ ] Text: Replace "Initialize Memory Bank"
 -   [ ] Webview: Memory Bank: always red "missing"
@@ -59,11 +62,12 @@
 
 ---
 
-**Notes (2025-05-11):**
-- The ruleset and `.mdc` file are now up-to-date; self-healing is robust and automatic.
-- The "Reset the rules" button in the webview now provides user feedback and error handling.
-- MCP tool usage is now enforced and documented in the rules.
+**Notes (2025-05-13):**
+- The codebase is now Cursor-first, with VS Code compatibility as a bonus.
+- Modularisation, webview overhaul, MCP CLI/stdio, and self-healing are complete.
+- Express removal is the last major refactor in progress.
+- Docs, rules, and packaging are up-to-date.
 
 For a detailed, step-by-step experimental plan to safely prototype advanced MCP features (chunked file access, metadata, planner tools), see [EXPERIMENTAL-MCP-PLAN.md](../experimental/EXPERIMENTAL-MCP-PLAN.md).
 
-_Last updated: 2025-05-11 🐹_
+_Last updated: 2025-05-13 🐹_
