+++ package.json
@@ -6,7 +6,7 @@
     "url": "https://github.com/sm-moshi/aimemory"
   },
   "description": "Easily manage AI context for your projects using memory bank technique",
-  "version": "0.2.4",
+  "version": "0.2.5",
   "engines": {
     "vscode": "^1.96.2"
   },
@@ -111,7 +111,7 @@
     "express": "^4.21.2",
     "zod": "^3.24.4"
   },
-  "packageManager": "pnpm@10.10.0",
+  "packageManager": "pnpm@10.11.0",
   "pnpm": {
     "overrides": {
       "jwa": "^2.0.0"
