+++ tsconfig.json
@@ -13,6 +13,7 @@
 		"skipLibCheck": true,
 		/* Additional Checks */
 		"noImplicitReturns": true, /* Report error when not all code paths in function return a value. */
+		"types": ["node", "mocha", "vscode"],
 		// "noFallthroughCasesInSwitch": true, /* Report errors for fallthrough cases in switch statement. */
 		// "noUnusedParameters": true,  /* Report errors on unused parameters. */
 	},
